+++ HugeFiles/Tests/JsonChunkerTests.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Runtime.Versioning;
 using System.Text;
 using HugeFiles.HugeFiles;
 using HugeFiles.Utils;
@@ -10,6 +11,46 @@ namespace HugeFiles.Tests
 {
     public class JsonChunkerTester
     {
+        private static (int ii, int tests_failed) TestOneChunk(int ii, int tests_failed, 
+                                                                string fname, int minChunk, int maxChunk,
+                                                                List<Chunk> correctChunks, JsonChunker chunker)
+        {
+            string failureMessage = $"While JsonChunking {fname} with minChunk={minChunk}, maxChunk={maxChunk}, ";
+            try
+            {
+                chunker.Parse();
+            }
+            catch (Exception ex)
+            {
+                ii++;
+                tests_failed++;
+                Npp.AddLine(failureMessage + $"got error:\r\n{ex}");
+                return (ii, tests_failed);
+            }
+            ii++;
+            int correctCount = correctChunks.Count;
+            int chunkCount = chunker.chunks.Count;
+            if (chunkCount != correctCount)
+            {
+                tests_failed++;
+                Npp.AddLine(failureMessage + $"expected {correctCount} chunks but got {chunkCount}");
+                return (ii, tests_failed);
+            }
+            for (int jj = 0; jj < chunkCount; jj++)
+            {
+                ii++;
+                Chunk gotChunk = chunker.chunks[jj];
+                Chunk correctChunk = correctChunks[jj];
+                if (gotChunk.start != correctChunk.start
+                    || gotChunk.end != correctChunk.end)
+                {
+                    tests_failed++;
+                    Npp.AddLine(failureMessage + $"expected chunk {jj} to start at char {correctChunk.start} and end at {correctChunk.end}, but instead started at {gotChunk.start} and ended at {gotChunk.end}");
+                }
+            }
+            return (ii, tests_failed);
+        }
+
         public static void Test()
         {
             int ii = 0;
@@ -92,6 +133,9 @@ namespace HugeFiles.Tests
             };
             string curdir = "plugins/HugeFiles/testfiles/";
             JsonChunker chunker = null;
+            /********
+             * TEST IF CHUNKER ADDS THE CORRECT NUMBER OF CHUNKS WITH BOUNDARIES IN THE RIGHT PLACES
+            *********/
             foreach ((int minChunk, int maxChunk, string fname,
                 List<Chunk> correctChunks) in testcases)
             {
@@ -100,40 +144,25 @@ namespace HugeFiles.Tests
                     chunker.Dispose();
                 }
                 chunker = new JsonChunker(curdir + fname, minChunk, maxChunk);
-                string failureMessage = $"While JsonChunking {fname} with minChunk={minChunk}, maxChunk={maxChunk}, ";
-                try
-                {
-                    chunker.Parse();
-                }
-                catch (Exception ex)
-                {
-                    ii++;
-                    tests_failed++;
-                    Npp.AddLine(failureMessage + $"got error:\r\n{ex}");
-                    continue;
-                }
-                ii++;
-                int correctCount = correctChunks.Count;
-                int chunkCount = chunker.chunks.Count;
-                if (chunkCount != correctCount)
-                {
-                    tests_failed++;
-                    Npp.AddLine(failureMessage + $"expected {correctCount} chunks but got {chunkCount}");
-                    continue;
-                }
-                for (int jj = 0; jj < chunkCount; jj++)
-                {
-                    ii++;
-                    Chunk gotChunk = chunker.chunks[jj];
-                    Chunk correctChunk = correctChunks[jj];
-                    if (gotChunk.start != correctChunk.start
-                        || gotChunk.end != correctChunk.end)
-                    {
-                        tests_failed++;
-                        Npp.AddLine(failureMessage + $"expected chunk {jj} to start at char {correctChunk.start} and end at {correctChunk.end}, but instead started at {gotChunk.start} and ended at {gotChunk.end}");
-                    }
-                }
+                (ii, tests_failed) = TestOneChunk(ii, tests_failed, fname, minChunk, maxChunk, correctChunks, chunker);
             }
+
+            /********
+             * TEST IF CHUNKER CHUNKS A FILE CORRECTLY EVEN AFTER SWITCHING BETWEEN FILES
+            *********/
+            int minch, maxch;
+            string fname_;
+            List<Chunk> corChunks;
+            (minch, maxch, fname_, corChunks) = testcases[0];
+            if (chunker != null)
+                chunker.Dispose();
+            chunker = new JsonChunker(curdir + fname_, minch, maxch);
+            (ii, tests_failed) = TestOneChunk(ii, tests_failed, fname_, minch, maxch, corChunks, chunker);
+            // now switch to a new file and test if it makes the right chunks
+            (minch, maxch, fname_, corChunks) = testcases[1];
+            chunker.Reset(",", minch, maxch);
+            chunker.ChooseNewFile(curdir + fname_);
+            (ii, tests_failed) = TestOneChunk(ii, tests_failed, fname_, minch, maxch, corChunks, chunker);
 
             Npp.AddLine($"Failed {tests_failed} tests.");
             Npp.AddLine($"Passed {ii - tests_failed} tests.");